requirements.txt
docker-compose==1.7.1
gitpython